+++ ExpandWorldPrefabs/service/data/DataHelper.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Service;
@@ -42,10 +43,7 @@ public class DataHelper
     var prefab = Utils.GetPrefabName(obj).GetStableHashCode();
     ZNetView.m_initZDO = ZDOMan.instance.CreateNewZDO(pos, prefab);
     if (data != null)
-    {
-      pars = data.InsertParameters(pars, ZNetView.m_initZDO);
-      data.Write(pars, ZNetView.m_initZDO);
-    }
+      data.Write([], ZNetView.m_initZDO);
     ZNetView.m_initZDO.m_rotation = rot.eulerAngles;
     ZNetView.m_initZDO.Type = view.m_type;
     ZNetView.m_initZDO.Distant = view.m_distant;
@@ -64,10 +62,7 @@ public class DataHelper
     var prefab = Utils.GetPrefabName(obj).GetStableHashCode();
     ZNetView.m_initZDO = ZDOMan.instance.CreateNewZDO(pos, prefab);
     if (data != null)
-    {
-      var pars = data.InsertParameters([], ZNetView.m_initZDO);
-      data.Write(pars, ZNetView.m_initZDO);
-    }
+      data.Write([], ZNetView.m_initZDO);
     ZNetView.m_initZDO.m_rotation = rot.eulerAngles;
     ZNetView.m_initZDO.Type = view.m_type;
     ZNetView.m_initZDO.Distant = view.m_distant;
@@ -122,5 +117,30 @@ public class DataHelper
   {
     var lower = key.ToLowerInvariant();
     return ZoneSystem.instance.m_globalKeysValues.FirstOrDefault(kvp => kvp.Key.ToLowerInvariant() == lower).Value ?? "0";
+  }
+
+  // Parameter value could be a value group, so that has to be resolved.
+  public static string ResolveValue(string value)
+  {
+    if (!value.StartsWith("<", StringComparison.OrdinalIgnoreCase)) return value;
+    if (!value.EndsWith(">", StringComparison.OrdinalIgnoreCase)) return value;
+    var sub = value.Substring(1, value.Length - 2);
+    if (TryGetValueFromGroup(sub, out var valueFromGroup))
+      return valueFromGroup;
+    return value;
+  }
+
+  public static bool TryGetValueFromGroup(string group, out string value)
+  {
+    var hash = group.ToLowerInvariant().GetStableHashCode();
+    if (!DataLoading.ValueGroups.ContainsKey(hash))
+    {
+      value = group;
+      return false;
+    }
+    var roll = UnityEngine.Random.Range(0, DataLoading.ValueGroups[hash].Count);
+    // Value from group could be another group, so yet another resolve is needed.
+    value = ResolveValue(DataLoading.ValueGroups[hash][roll]);
+    return true;
   }
 }