+++ ExpandWorldPrefabs/service/data/values/BoolValue.cs
@@ -1,23 +1,20 @@
-
-using System.Collections.Generic;
-
 namespace Data;
 
 public class BoolValue(string[] values) : AnyValue(values), IBoolValue
 {
-  public int? GetInt(Dictionary<string, string> pars)
+  public int? GetInt(Pars pars)
   {
     var value = GetValue(pars);
     if (value == null) return null;
     return value == "true" ? 1 : 0;
   }
-  public bool? GetBool(Dictionary<string, string> pars)
+  public bool? GetBool(Pars pars)
   {
     var value = GetValue(pars);
     if (value == null) return null;
     return value == "true";
   }
-  public bool? Match(Dictionary<string, string> pars, bool value)
+  public bool? Match(Pars pars, bool value)
   {
 
     // If all values are null, default to a match.
@@ -39,14 +36,14 @@ public class SimpleBoolValue(bool value) : IBoolValue
 {
   private readonly bool Value = value;
 
-  public int? GetInt(Dictionary<string, string> pars) => Value ? 1 : 0;
-  public bool? GetBool(Dictionary<string, string> pars) => Value;
-  public bool? Match(Dictionary<string, string> pars, bool value) => Value == value;
+  public int? GetInt(Pars pars) => Value ? 1 : 0;
+  public bool? GetBool(Pars pars) => Value;
+  public bool? Match(Pars pars, bool value) => Value == value;
 }
 
 public interface IBoolValue
 {
-  int? GetInt(Dictionary<string, string> pars);
-  bool? GetBool(Dictionary<string, string> pars);
-  bool? Match(Dictionary<string, string> pars, bool value);
+  int? GetInt(Pars pars);
+  bool? GetBool(Pars pars);
+  bool? Match(Pars pars, bool value);
 }