+++ ExpandWorldPrefabs/service/data/values/HashValue.cs
@@ -1,13 +1,11 @@
-
-using System.Collections.Generic;
 using System.Linq;
 
 namespace Data;
 
 public class HashValue(string[] values) : AnyValue(values), IHashValue
 {
-  public int? Get(Dictionary<string, string> pars) => GetValue(pars)?.GetStableHashCode();
-  public bool? Match(Dictionary<string, string> pars, int value)
+  public int? Get(Pars pars) => GetValue(pars)?.GetStableHashCode();
+  public bool? Match(Pars pars, int value)
   {
     var values = GetAllValues(pars);
     if (values.Length == 0) return null;
@@ -18,11 +16,11 @@ public class SimpleHashValue(string value) : IHashValue
 {
   private readonly int Value = value.GetStableHashCode();
 
-  public int? Get(Dictionary<string, string> pars) => Value;
-  public bool? Match(Dictionary<string, string> pars, int value) => Value == value;
+  public int? Get(Pars pars) => Value;
+  public bool? Match(Pars pars, int value) => Value == value;
 }
 public interface IHashValue
 {
-  int? Get(Dictionary<string, string> pars);
-  bool? Match(Dictionary<string, string> pars, int value);
+  int? Get(Pars pars);
+  bool? Match(Pars pars, int value);
 }