+++ ExpandWorldPrefabs/service/data/values/IntValue.cs
@@ -1,14 +1,11 @@
-
-using System.Collections.Generic;
 using System.Globalization;
-using Service;
 using UnityEngine;
 
 namespace Data;
 
 public class IntValue(string[] values) : AnyValue(values), IIntValue
 {
-  public int? Get(Dictionary<string, string> pars)
+  public int? Get(Pars pars)
   {
     var value = GetValue(pars);
     if (value == null)
@@ -42,7 +39,7 @@ public class IntValue(string[] values) : AnyValue(values), IIntValue
       return roll;
     return Calculator.EvaluateInt(split[3].Replace("<value>", roll.ToString()));
   }
-  public bool? Match(Dictionary<string, string> pars, int value)
+  public bool? Match(Pars pars, int value)
   {
     // If all values are null, default to a match.
     var allNull = true;
@@ -127,11 +124,11 @@ public class IntValue(string[] values) : AnyValue(values), IIntValue
 public class SimpleIntValue(int value) : IIntValue
 {
   private readonly int Value = value;
-  public int? Get(Dictionary<string, string> pars) => Value;
-  public bool? Match(Dictionary<string, string> pars, int value) => Value == value;
+  public int? Get(Pars pars) => Value;
+  public bool? Match(Pars pars, int value) => Value == value;
 }
 public interface IIntValue
 {
-  int? Get(Dictionary<string, string> pars);
-  bool? Match(Dictionary<string, string> pars, int value);
+  int? Get(Pars pars);
+  bool? Match(Pars pars, int value);
 }