+++ ExpandWorldPrefabs/service/data/values/QuaternionValue.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Linq;
 using Service;
 using UnityEngine;
@@ -7,8 +6,8 @@ namespace Data;
 
 public class QuaternionValue(string[] values) : AnyValue(values), IQuaternionValue
 {
-  public Quaternion? Get(Dictionary<string, string> pars) => Parse.AngleYXZNull(GetValue(pars));
-  public bool? Match(Dictionary<string, string> pars, Quaternion value)
+  public Quaternion? Get(Pars pars) => Parse.AngleYXZNull(GetValue(pars));
+  public bool? Match(Pars pars, Quaternion value)
   {
     var values = GetAllValues(pars);
     if (values.Length == 0) return null;
@@ -19,12 +18,12 @@ public class QuaternionValue(string[] values) : AnyValue(values), IQuaternionVal
 public class SimpleQuaternionValue(Quaternion value) : IQuaternionValue
 {
   private readonly Quaternion Value = value;
-  public Quaternion? Get(Dictionary<string, string> pars) => Value;
-  public bool? Match(Dictionary<string, string> pars, Quaternion value) => Value == value;
+  public Quaternion? Get(Pars pars) => Value;
+  public bool? Match(Pars pars, Quaternion value) => Value == value;
 }
 
 public interface IQuaternionValue
 {
-  Quaternion? Get(Dictionary<string, string> pars);
-  bool? Match(Dictionary<string, string> pars, Quaternion value);
+  Quaternion? Get(Pars pars);
+  bool? Match(Pars pars, Quaternion value);
 }