+++ ExpandWorldPrefabs/service/data/values/Vector3Value.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Linq;
 using Service;
 using UnityEngine;
@@ -7,8 +6,8 @@ namespace Data;
 
 public class Vector3Value(string[] values) : AnyValue(values), IVector3Value
 {
-  public Vector3? Get(Dictionary<string, string> pars) => Parse.VectorXZYNull(GetValue(pars));
-  public bool? Match(Dictionary<string, string> pars, Vector3 value)
+  public Vector3? Get(Pars pars) => Parse.VectorXZYNull(GetValue(pars));
+  public bool? Match(Pars pars, Vector3 value)
   {
     var values = GetAllValues(pars);
     if (values.Length == 0) return null;
@@ -19,12 +18,12 @@ public class Vector3Value(string[] values) : AnyValue(values), IVector3Value
 public class SimpleVector3Value(Vector3 value) : IVector3Value
 {
   private readonly Vector3 Value = value;
-  public Vector3? Get(Dictionary<string, string> pars) => Value;
-  public bool? Match(Dictionary<string, string> pars, Vector3 value) => Value == value;
+  public Vector3? Get(Pars pars) => Value;
+  public bool? Match(Pars pars, Vector3 value) => Value == value;
 }
 
 public interface IVector3Value
 {
-  Vector3? Get(Dictionary<string, string> pars);
-  bool? Match(Dictionary<string, string> pars, Vector3 value);
+  Vector3? Get(Pars pars);
+  bool? Match(Pars pars, Vector3 value);
 }